+++ views/dogs.py
@@ -1,6 +1,6 @@
 # encoding: utf-8
 from flask import jsonify, request
-from flask_login import login_required
+from flask_login import login_required, current_user
 from project.app import db
 from project.models import Dog
 from utils.rest import RestView
@@ -9,6 +9,7 @@ from utils.rest import RestView
 class DogAPI(RestView):
     schema = 'Dog'
 
+    @login_required
     def get(self, id):
         page = request.args.get('page') or 1
         if id is None:
@@ -26,6 +27,7 @@ class DogAPI(RestView):
     def post(self):
         data = request.get_json()
         dog = Dog(**data)
+        dog.user = current_user
         db.session.add(dog)
         db.session.commit()
         return self.make_response('Dog created successfully.', 201)
@@ -40,6 +42,9 @@ class DogAPI(RestView):
     @login_required
     def delete(self, id):
         dog = Dog.query.get(id)
-        db.session.delete(dog)
-        db.session.commit()
-        return self.make_response('Dog deleted successfully.')
+        if dog:
+            db.session.delete(dog)
+            db.session.commit()
+            return self.make_response('Dog deleted successfully.')
+        else:
+            return self.make_response('Dog not found.', 404)
